backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
pandas==3.0.2
openpyxl==3.1.5
lxml==6.0.4
python-dotenv==1.2.2
gunicorn==25.3.0
paramiko==4.0.0
pytest==9.0.3